fixed taskbar icon not showing up
