+++ src/main/java/com/treasure/hunt/Main.java
@@ -4,13 +4,30 @@ import com.sun.javafx.application.LauncherImpl;
 import com.treasure.hunt.view.SplashScreenLoader;
 import lombok.extern.slf4j.Slf4j;
 
+import java.awt.*;
+
 /**
  * @author axel12
  */
 @Slf4j
 public class Main {
+
+    static {
+        System.getProperties().setProperty("-Xdock:name", "TreasureHunt");
+    }
+
     public static void main(String[] args) {
         log.info("Starting Javafx UI application");
+        setTaskBarIcon();
         LauncherImpl.launchApplication(JavaFxApplication.class, SplashScreenLoader.class, args);
+    }
+
+    private static void setTaskBarIcon() {
+        Image image = Toolkit.getDefaultToolkit().getImage(JavaFxApplication.class.getResource("/images/icon.png"));
+        try {
+            Taskbar.getTaskbar().setIconImage(image);
+        } catch (Exception e) {
+            log.info("This platform seems to not support taskbar icon image");
+        }
     }
 }
